crypto_quant_v16/requirements.txt
ccxt>=4.5.0
numpy>=2.1.0
pandas>=2.2.0
panel>=1.4.0
plotly>=5.24.0
streamlit>=1.40.0
ta>=0.11.0
scikit-learn>=1.6.0
matplotlib>=3.9.0
schedule>=1.2.0
yfinance>=0.2.50
optuna>=4.0.0
python-dotenv>=1.0.0
requests>=2.32.0
scipy>=1.14.0
